+++ app/src/main/AndroidManifest.xml
@@ -3,10 +3,13 @@
     package="com.nicholas.rutherford.track.your.shot">
 
     <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"
+        android:maxSdkVersion="32" />
     <uses-permission android:name="android.permission.READ_MEDIA_IMAGES" />
     <uses-permission android:name="android.permission.CAMERA" />
 
     <uses-feature android:name="android.hardware.camera" />
+
 
     <application
         android:name=".MyApplication"
