+++ feature/players/src/main/java/com/nicholas/rutherford/track/your/shot/feature/players/createplayer/CreatePlayerParams.kt
@@ -10,5 +10,6 @@ data class CreatePlayerParams(
     val onImageUploadClicked: (uri: Uri?) -> Unit,
     val onCreatePlayerClicked: (uri: Uri?) -> Unit,
     val permissionNotGrantedForCameraAlert: () -> Unit,
+    val permissionNotGrantedForReadMediaOrExternalStorageAlert: () -> Unit,
     val onSelectedCreateEditImageOption: (uri: String) -> CreateEditImageOption
 )
