+++ feature/players/src/main/java/com/nicholas/rutherford/track/your/shot/feature/players/createplayer/ext/UploadPlayerImageContent.kt
@@ -1,0 +1,60 @@
+package com.nicholas.rutherford.track.your.shot.feature.players.createplayer.ext
+
+import android.net.Uri
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.ExperimentalMaterialApi
+import androidx.compose.material.Icon
+import androidx.compose.material.ModalBottomSheetState
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AddAPhoto
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.layout.ContentScale
+import androidx.compose.ui.unit.dp
+import coil.compose.AsyncImage
+import com.nicholas.rutherford.track.your.shot.feature.players.createplayer.CreatePlayerParams
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.launch
+
+@OptIn(ExperimentalMaterialApi::class)
+@Composable
+fun UploadPlayerImageContent(
+    hasUploadedImage: Boolean,
+    scope: CoroutineScope,
+    bottomState: ModalBottomSheetState,
+    createPlayerParams: CreatePlayerParams,
+    imageUri: Uri?
+) {
+    Column(modifier = Modifier.fillMaxWidth()) {
+        if (hasUploadedImage && imageUri != null) {
+            AsyncImage(
+                model = imageUri,
+                contentDescription = null,
+                modifier = Modifier
+                    .size(90.dp)
+                    .clickable {
+                        createPlayerParams.onImageUploadClicked.invoke(imageUri)
+                        scope.launch { bottomState.show() }
+                    }
+                    .clip(CircleShape),
+                contentScale = ContentScale.Crop
+            )
+        } else {
+            Icon(
+                imageVector = Icons.Default.AddAPhoto,
+                contentDescription = "Add a photo icon",
+                modifier = Modifier
+                    .size(48.dp)
+                    .clickable {
+                        createPlayerParams.onImageUploadClicked.invoke(imageUri)
+                        scope.launch { bottomState.show() }
+                    }
+            )
+        }
+    }
+}
